Restore the Fselect_window call in gui_consider_frame_title.
This fixes bug #55412.  The call to Fselect_window was removed on 2021-03-21
in the commit "Prevent open minibuffers getting lost when their frame gets
deleted".  This call is actually needed to determine current elements of the
pertinent window and frame when these are used in the frame title.

* src/frame.c (do_switch_frame): When the selected window in the target frame
is the mini-window, switch away from this window unless there is a valid
minibuffer there.

* src/frame.h (struct frame): Add an incidental comment about the element
selected_window.

* src/minibuf.c (move_minibuffers_onto_frame): No longer set the selected
window of the source frame.  This action is now performed later, on returning
to that frame, in do_switch_frame when needed (see above).

* src/xdisp.c (gui_consider_frame_title): Replace the Fselect_window call and
associated ancillary settings.