Remove obscure, obsolete code from do_switch_frame
This is relevant for bug #56305, and might solve that bug.  The code being
removed went into Emacs between 1992 and 1994, and looks to have been a
workaround for switching frames, before the command 'other-frame' had been
written.  Nowadays, that code has harmful effects, causing frames' focus to be
redirected at random, sometimes back to the frame itself.

* src/frame.c (do_switch_frame): Remove 53 lines of code.