Fix horizontal off-by-one error
Delay increasing width used by columns until we know that we have to
do so because we have determined that there is enough room to add an
additional column and a space between the last two columns.

If we don't do that, then we can easily get an off-by-one error.  If
docstrings are shown and the window is narrow, then it is likely that
we end up using the maximal width.  If we then add one to the actual
width and later compare that again with the maximal width, then that
is too width.